refactor(adr): nest `domain` command under `adr`
Move `archgate domain ...` to `archgate adr domain ...` so the command
lives next to the other ADR subcommands (create, list, show, update) it
pairs with. No behavior changes — only the invocation path.
