+++ src/commands/adr/index.ts
@@ -1,6 +1,7 @@
 import type { Command } from "@commander-js/extra-typings";
 
 import { registerAdrCreateCommand } from "./create";
+import { registerDomainCommand } from "./domain/index";
 import { registerAdrListCommand } from "./list";
 import { registerAdrShowCommand } from "./show";
 import { registerAdrUpdateCommand } from "./update";
@@ -14,4 +15,5 @@ export function registerAdrCommand(program: Command) {
   registerAdrListCommand(adr);
   registerAdrShowCommand(adr);
   registerAdrUpdateCommand(adr);
+  registerDomainCommand(adr);
 }
